+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 5.10.3
 * Fixed button positioning issue when using the mouse to point at menu items
+* Fixed bug in customizing highlightTarget
 
 ## Version 5.10.2
 * Fixed issue with popup button not being hidden in some cases with a CSS fixed header